% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tests.R
\name{are_false}
\alias{are_false}
\title{Which elements in a list are FALSE?}
\usage{
are_false(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
A logical vector.
}
\description{
Which elements in a list are FALSE?
}
\examples{
are_false(list(FALSE,TRUE,TRUE))
}
\seealso{
Other Tests: \code{\link{all_empty}},
  \code{\link{are_empty}}, \code{\link{is.empty}};
  \code{\link{all_scalar}}, \code{\link{are_scalar}},
  \code{\link{is.scalar}}, \code{\link{is.string}};
  \code{\link{are_null}}; \code{\link{are_true}};
  \code{\link{has_command}}; \code{\link{is.installed}}
}

