% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tests.R
\name{are_null}
\alias{are_null}
\title{Which elements in a list are NULL?}
\usage{
are_null(x)
}
\arguments{
\item{x}{An object to test.}

\item{A}{logical vector.}
}
\description{
Which elements in a list are NULL?
}
\examples{
are_null(list(1, NULL, 3))
}
\seealso{
Other Tests: \code{\link{all_empty}},
  \code{\link{are_empty}}, \code{\link{is.empty}};
  \code{\link{all_scalar}}, \code{\link{are_scalar}},
  \code{\link{is.scalar}}, \code{\link{is.string}};
  \code{\link{are_false}}; \code{\link{are_true}};
  \code{\link{has_command}}; \code{\link{is.installed}}
}

