% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{are_true}
\alias{are_true}
\title{Which elements in a list are TRUE?}
\usage{
are_true(x)
}
\arguments{
\item{x}{An object to test.}

\item{A}{logical vector.}
}
\description{
Which elements in a list are TRUE?
}
\examples{
are_true(list(FALSE, TRUE, TRUE))
}
\seealso{
Other Tests: \code{\link{all_empty}},
  \code{\link{are_empty}}, \code{\link{is.empty}};
  \code{\link{all_scalar}}, \code{\link{are_scalar}},
  \code{\link{is.scalar}}, \code{\link{is.string}};
  \code{\link{are_false}}; \code{\link{are_null}};
  \code{\link{has_command}}; \code{\link{is.installed}}
}

