% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{blanks}
\alias{blanks}
\title{Create blank strings with a given number of characters.}
\usage{
blanks(n)
}
\arguments{
\item{n}{A numeric vector (number of times to duplicate).}
}
\value{
A character vector.
}
\description{
Create blank strings with a given number of characters.
}
\examples{
blanks(10)
}
\seealso{
Other string-utils: \code{\link{dup}}; \code{\link{pad}};
  \code{\link{replace_ext}}; \code{\link{split_path}};
  \code{\link{strip_ext}}; \code{\link{strsplitN}};
  \code{\link{trim}}; \code{\link{wrap}}
}

