% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{compact}
\alias{compact}
\alias{compactAll}
\alias{compactChar}
\alias{compactNA}
\title{Filter "empty" elements from a list.}
\usage{
compact(x)

compactChar(x)

compactNA(x)

compactAll(x)
}
\arguments{
\item{x}{A vector or list.}
}
\value{
A vector or list with empty elements filtered out.
}
\description{
\code{compact} filters \code{NULL} elments from a list; \code{compactChar} filters
empty strings (\code{""}) from a list (or vector); \code{compactNA} filters
\code{NA}s and  \code{compactAll} filters all of the above.
}
\examples{
l <- list(a=1, b=NULL, c=NA)
compact(l)
## $a
## [1] 1
##
## $c
## [1] NA
compactNA(l)
## $a
## [1] 1
##
## $b
## NULL
}

