% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createProject}
\alias{createProject}
\title{Create a modified \href{http://projecttemplate.net/getting_started.html}{ProjectTemplate}
project.}
\usage{
createProject(project = "myProject", path = getOption("gsmisc.proj"),
  merge_strategy = c("require.empty", "allow.non.conflict"), open = TRUE,
  packrat = FALSE)
}
\arguments{
\item{project}{A character vector with the name of the project directory.}

\item{path}{The base directory of the project.}

\item{merge_strategy}{What should happen if the target directory exists and
is not empty?
If \code{"force.empty"}, the target directory must be empty;
if \code{"allow.non.conflict"}, the method succeeds if no files or
directories with the same name exist in the target directory.}

\item{open}{If \code{TRUE}, open the newly created project in RStudio.}

\item{packrat}{If \code{TRUE}, create a \href{http://rstudio.github.io/packrat/walkthrough.html}{packrat} project.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
Create a modified \href{http://projecttemplate.net/getting_started.html}{ProjectTemplate}
project.
}
\details{
If the target directory does not exist, it is created.  Otherwise,
  it can only contain files and directories allowed by the merge strategy.
}
\note{
This function is slightly modified code taken from the
  \code{\link[ProjectTemplate]{create.project}} function in John Myles White's
  \code{ProjectTemplate} package to allow me using my custom directory structure.
}

