% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite-utils.R
\name{db_disconnect}
\alias{db_disconnect}
\title{Disconnect from an SQLite database.}
\usage{
db_disconnect(...)
}
\arguments{
\item{...}{Connection objects.}
}
\description{
Disconnect from an SQLite database.
}
\keyword{internal}
