% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%||\%}
\alias{\%||\%}
\alias{\%|\%}
\alias{\%|na|\%}
\alias{\%|NA|\%}
\alias{\%|null|\%}
\title{A null coalescing operator.}
\usage{
a \%||\% b

a \%|\% b

a \%|na|\% b

a \%|NA|\% b

a \%|null|\% b
}
\arguments{
\item{a}{Some value.}

\item{b}{Some default value.}
}
\value{
\code{a}, if \code{a} is not \code{NULL}, not an empty
vector, not an empty string, or not \code{NA}.
}
\description{
If the LHS eveluates to \code{NULL}, a zero-length vector, or
an empty string, return the RHS, otherwise return the LHS.
}
\details{
\code{\%|\%} is a vectorized version of \code{\%||\%}

\code{\%|na|\%} returns the RHS if the LHS evaluates to \code{NA}

\code{\%|NA|\%} is a vectorized version of \code{\%|na|\%}

\code{\%|null|\%} explicitly works only on \code{NULL} but
not empty strings and vectors.
}
\examples{
a1 <- "some value"
a2 <- ""
b <- "default value"
a1 \%||\% b
a2 \%||\% b
}
\keyword{utilities}
