% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{Dot_every}
\alias{Dot_every}
\title{Print a dot ever nth function call}
\usage{
Dot_every(n, fn)
}
\arguments{
\item{n}{when to print a dot}

\item{fn}{function call}
}
\description{
Print a dot ever nth function call
}
\seealso{
Other Functionals: \code{\link{And}}, \code{\link{Call}},
  \code{\link{Compose}}, \code{\link{Delay_by}},
  \code{\link{FMap}}, \code{\link{Fail_with}},
  \code{\link{Log_to}}, \code{\link{Maybe}},
  \code{\link{Partial}}
}

