% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-utils.R
\name{dup}
\alias{dup}
\title{Duplicate a character string n times.}
\usage{
dup(x, n)
}
\arguments{
\item{x}{Input character string.}

\item{n}{A numeric vector (number of times to duplicate).}
}
\value{
A character string.
}
\description{
Duplicate a character string n times.
}
\examples{
dup("#", 1:10)
}
\seealso{
Other string-utils: \code{\link{blanks}},
  \code{\link{pad}}, \code{\link{replace_ext}},
  \code{\link{split_path}}, \code{\link{strip_ext}},
  \code{\link{strsplitN}}, \code{\link{trim}},
  \code{\link{wrap}}
}

