% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{\%<~\%}
\alias{\%<~\%}
\alias{\%<~f\%}
\title{A pattern extraction operator.}
\usage{
x \%<~\% re

x \%<~f\% re
}
\arguments{
\item{x}{A character vector to match \code{re} against.}

\item{re}{A regular expression.}
}
\value{
The matching elements in \code{x}.
}
\description{
Extract matching elements from a character vector.
}
\seealso{
Other Operators: \code{\link{\%n~\%}},
  \code{\link{\%n~f\%}}; \code{\link{\%~\%}},
  \code{\link{\%~f\%}}

Other Operators: \code{\link{\%n~\%}},
  \code{\link{\%n~f\%}}; \code{\link{\%~\%}},
  \code{\link{\%~f\%}}
}

