% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Fail_with}
\alias{Fail_with}
\title{Fail with a default value}
\usage{
Fail_with(default = NULL, fn, verbose = TRUE)
}
\arguments{
\item{default}{Default value.}

\item{fn}{A function}

\item{verbose}{Show error message}
}
\description{
Fail with a default value
}
\seealso{
Other Functionals: \code{\link{And}}, \code{\link{Or}};
  \code{\link{Call}}; \code{\link{Compose}},
  \code{\link{Sequence}}, \code{\link{\%.\%}};
  \code{\link{Delay_by}}; \code{\link{Dot_every}};
  \code{\link{FMap}}; \code{\link{Log_to}};
  \code{\link{Maybe}}; \code{\link{Partial}}
}

