% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{google}
\alias{google}
\title{Use Google Search from the R command line.}
\usage{
google(query = "", n = 10)
}
\arguments{
\item{query}{Search query}

\item{n}{Max number of results returned.}
}
\value{
Prints the URLs returned by Google Search to screen and invisibly
returns a character vector of the URLs.
}
\description{
Use Google Search from the R command line.
}
\examples{
r <- google("use google from command line")
\dontrun{
browseURL(r[1])
}
}

