% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/operators.R
\name{\%~\%}
\alias{\%~\%}
\title{A simple pattern matching operator.}
\usage{
x \%~\% re
}
\arguments{
\item{x}{A character vector.}

\item{pattern}{A regular expression.}
}
\value{
Matching elements in \code{target}
}
\description{
A simple pattern matching operator.
}

