% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{has_command}
\alias{has_command}
\title{Test if an external executable is available.}
\usage{
has_command(cmd, msg = "")
}
\arguments{
\item{cmd}{An exececutable to test for.}

\item{msg}{Additional message if the test fails.}
}
\value{
A logical.
}
\description{
Uses \code{\link{Sys.which}} internally, so it should work
on Windows and Unix.alikes.
}
\seealso{
Other Tests: \code{\link{are_false}},
  \code{\link{are_null}}, \code{\link{are_true}},
  \code{\link{is.empty}}, \code{\link{is.installed}},
  \code{\link{is.scalar}}
}
