% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_packages}
\alias{install_packages}
\alias{update_packages}
\title{Install or update packages.}
\usage{
install_packages(pkgs, destdir = getOption("gsmisc.pkgs"), update = FALSE,
  ...)

update_packages(destdir = getOption("gsmisc.pkgs"))
}
\arguments{
\item{pkgs}{Packages}

\item{destdir}{Where to put the source code.}

\item{update}{call \code{\link{update_packages}}.}

\item{...}{Further arguments passed on to \code{\link{biocLite}}.}
}
\value{
Called for its side effects.
}
\description{
\code{install_packages} is a wrapper around \code{\link{biocLite}}.
In addition to the standard functionality provided \code{biocLite},
this function also downloads and untars the source files to \code{destdir}.
}
\details{
\code{update_packages} checks for outdated packages, save the source to
\code{destdir}, extract the packages in \code{destdir}, and installs them.
}
\seealso{
Other Installers: \code{\link{update_github}}
}
