% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_pkg_set}
\alias{install_pkg_set}
\title{Install different sets of packages}
\usage{
install_pkg_set(which = "base", ...)
}
\arguments{
\item{which}{Which set of packages do we want to install.}
}
\description{
Different sets of packages:
\code{base}, \code{datatools}, \code{hadleyverse}, \code{stats},
\code{phlyo}, \code{bioc}.
}
