% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{is.empty}
\alias{is.empty}
\alias{are_empty}
\alias{all_empty}
\title{Is an object empty?}
\usage{
is.empty(x)

are_empty(x)

all_empty(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
A logical.
}
\description{
Which elements in a list are empty?
Are all elements in a list empty?
}
\examples{
is.empty(NULL)
is.empty(numeric())
is.empty(list())
is.empty("")
are_empty(list(1, NULL, 3, NA))
all_empty(NULL)
all_empty(list(NULL, NULL, character()))
}
\seealso{
Other Tests: \code{\link{are_false}},
  \code{\link{are_null}}, \code{\link{are_true}},
  \code{\link{has_command}}, \code{\link{is.installed}},
  \code{\link{is.scalar}}
}
