% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{is.installed}
\alias{is.installed}
\title{Is an R Packages installed?}
\usage{
is.installed(pkg)
}
\arguments{
\item{pkg}{Package name as character string.}
}
\value{
A logical.
}
\description{
Is an R Packages installed?
}
\examples{
is.installed("methods")
}
\seealso{
Other Tests: \code{\link{are_false}},
  \code{\link{are_null}}, \code{\link{are_true}},
  \code{\link{has_command}}, \code{\link{is.empty}},
  \code{\link{is.scalar}}
}

