% Generated by roxygen2 (4.0.2): do not edit by hand
\name{\%~\%}
\alias{\%~\%}
\alias{\%~f\%}
\title{A pattern matching operator.}
\usage{
x \%~\% re

x \%~f\% re
}
\arguments{
\item{x}{A character vector to match \code{re} against.}

\item{re}{A regular expression.}
}
\value{
A logical vector of matching elements in \code{x}.
}
\description{
A pattern matching operator.
}
\seealso{
Other Operators: \code{\link{\%<~\%}},
  \code{\link{\%<~f\%}}; \code{\link{\%n~\%}},
  \code{\link{\%n~f\%}}
}

