% Generated by roxygen2 (4.0.2): do not edit by hand
\name{\%ni\%}
\alias{\%ni\%}
\title{Reverse Value Matching}
\usage{
x \%ni\% y
}
\arguments{
\item{x}{The values to be matched.}

\item{y}{The values to \emph{not} be matched against.}
}
\value{
A logical vector.
}
\description{
Negation of the \code{\link{\%in\%}} operator.
}
\examples{
1:5 \%ni\% c(3,4,5)
}
\keyword{utilities}

