% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{nunique}
\alias{nunique}
\title{Number of unique elements in a vector.}
\usage{
nunique(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{passed to \code{\link{unique}}}
}
\description{
A wrapper around \code{length(unique(x))}
}

