% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pad}
\alias{pad}
\title{Pad a string.}
\usage{
pad(x, n = 10, where = "left", pad = " ")
}
\arguments{
\item{x}{Input character vector.}

\item{n}{Pad \code{x} to this (minimum) width.}

\item{where}{Side where the padding is added.}

\item{pad}{Padding character.}
}
\value{
A character vector.
}
\description{
Pad a string.
}
\seealso{
Other string-utils: \code{\link{blanks}};
  \code{\link{dup}}; \code{\link{replace_ext}};
  \code{\link{split_path}}; \code{\link{strip_ext}};
  \code{\link{strsplitN}}; \code{\link{trim}};
  \code{\link{wrap}}
}

