% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIRanges.R
\name{plotIRanges}
\alias{plotIRanges}
\title{Plot IRanges}
\source{
Modified after https://github.com/vsbuffalo/bds-files/blob/master/chapter-09-working-with-range-data/plot-ranges.R
}
\usage{
plotIRanges(..., sep = 0.5, height = 1, set_breaks = TRUE,
  labcol = "grey", names = NULL, cov = FALSE, clear = FALSE,
  disjoint = NULL, color = NULL)
}
\arguments{
\item{...}{One or more IRanges objects.}

\item{sep}{Vertical space between ranges.}

\item{height}{Height of bars.}

\item{set_breaks}{Breaks on x-axis.}

\item{labcol}{Colour of range labels.}

\item{names}{Names of ranges.}

\item{cov}{Coverage}

\item{clear}{Clear previous plot.}

\item{disjoint}{Manually assigned bins.}

\item{color}{Colour.}
}
\value{
A plot object.
}
\description{
Plot IRanges
}
\examples{
###
}
\author{
Vince Buffalo
}
