% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{replace_ext}
\alias{replace_ext}
\title{Replace file extensions.}
\usage{
replace_ext(file, replacement = "", sep = "\\\\.", level = 0)
}
\arguments{
\item{file}{The file name(s).}

\item{replacement}{replacement extension}

\item{sep}{Specifies the seperator character (default ".").}

\item{level}{How many extensions should be stripped.
The default (0) strips all, 1 strips the last one, 2 strips the last two,
and so on.}
}
\description{
Replace file extensions.
}
\seealso{
Other string-utils: \code{\link{blanks}};
  \code{\link{dup}}; \code{\link{pad}};
  \code{\link{split_path}}; \code{\link{strip_ext}};
  \code{\link{strsplitN}}; \code{\link{trim}};
  \code{\link{wrap}}
}

