% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tanken.R
\name{tanken_eingabe}
\alias{tanken_abruf}
\alias{tanken_eingabe}
\title{Eintragen und Abrufen von Daten in die tanken.db}
\usage{
tanken_eingabe(datum, tankstelle, adresse, kraftstoff = c("E10", "E05",
  "SUP"), menge, betrag, kilometerstand, initialisiern = FALSE)

tanken_abruf()
}
\arguments{
\item{datum}{Wann wurde getankt <dd.mm.yyyy>.}

\item{tankstelle}{Name der Tankstelle. (Aral, OMV, star, usw.)}

\item{adresse}{Adresse der Tankstelle. (zB. Stadtrodaer Straße 11, D-07749 Jena)}

\item{kraftstoff}{Eins von "E10", "E05", oder "SUP".}

\item{menge}{Getankte Kraftstoffmenge (in Liter).}

\item{betrag}{Gesamtbetrag in Euro.}

\item{kilometerstand}{Aktueller Gesamtkilometerstand.}

\item{initialisiern}{Falls \code{TRUE} wird die Datenbank initialisiert.}
}
\value{
Trägt übergebene Daten in die tanken.db ein.
}
\description{
Eintragen und Abrufen von Daten in die tanken.db
}
\examples{
\dontrun{
tanken_eingabe(initialisiern = TRUE)
tanken_eingabe(datum = "19.09.2014", tankstelle = "Aral",
               adresse = "Stadtrodaer Straße 11, 07749 Jena",
               kraftstoff = "E10", menge = 51.39, betrag = 76.01,
               kilometerstand = 433.8)
tanken_abruf()
}
}

