% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/string-utils.R
\name{trim}
\alias{trim}
\title{Trim elements}
\usage{
trim(x, trim = "\\\\s+")
}
\arguments{
\item{x}{A vector.}

\item{trim}{A pattern to trim from vector elements}
}
\value{
A vector with trimmed strings
}
\description{
Trim elements
}
\examples{
x <- " abc\\n\\t"
trim(x)
}
\seealso{
Other string-utils: \code{\link{blanks}};
  \code{\link{dup}}; \code{\link{pad}};
  \code{\link{replace_ext}}; \code{\link{split_path}};
  \code{\link{strip_ext}}; \code{\link{strsplitN}};
  \code{\link{wrap}}
}

