% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/install.R
\name{upgrade_bioc}
\alias{upgrade_bioc}
\title{Upgrade Bioconductor to the latest available version.}
\usage{
upgrade_bioc()
}
\description{
Downloads the latest version of the BiocInstaller package,
and upgrades all currently installed packages to the latest
repositories for this version of R.
}
\details{
Wraps \code{biocLite("BiocUpgrade")}
}

