% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{usplit}
\alias{usplit}
\title{unlist(strsplit(x, split, ...))}
\usage{
usplit(x, split, ...)
}
\arguments{
\item{\dots}{Arguments passed on to \code{\link{strsplit}}.}

\item{x}{Character vector to be split.}

\item{split}{The regexp to split on.}
}
\description{
unlist(strsplit(x, split, ...))
}
\examples{
usplit("a.b.c", ".", fixed = TRUE)
}
