% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{with}
\alias{with}
\alias{with_cpp11}
\alias{with_localtime}
\title{Execute code in an temporarily different environment.}
\usage{
with_cpp11(code)

with_localtime(new, code)
}
\arguments{
\item{code}{Code to execute.}

\item{new}{New value.}
}
\description{
\itemize{
  \item \code{with_cpp11}: temporarily set -std=c++11.
  \item \code{with_localtime}: temporarily change locale time.
  }
}
\examples{
\dontrun{
gsmisc <- as.package("~/R/Projects/Devel/gsmisc")
with_cpp11(devtools::build(gsmisc))
}
}

