% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{are_false}
\alias{are_false}
\title{Which elements in a list are FALSE?}
\usage{
are_false(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
A logical vector.
}
\description{
Which elements in a list are FALSE?
}
\examples{
are_false(list(FALSE,TRUE,TRUE))
}
\seealso{
Other Tests: \code{\link{are_null}},
  \code{\link{are_true}}, \code{\link{has_command}},
  \code{\link{is.empty}}, \code{\link{is.installed}},
  \code{\link{is.scalar}}
}
