% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{Call}
\alias{Call}
\title{Call a function with arguments provided individually.}
\usage{
Call(fn, ...)
}
\arguments{
\item{fn}{The function to call.}

\item{\dots}{Arguments to function \code{fn}.}
}
\value{
The return value of the function call.
}
\description{
Call a function with arguments provided individually.
}
\examples{
funs <- list("mean", "sd", "var")
sapply(funs, Call, 1:100)

## invoke an anonymous function
Call(function(a, b) a*b, 3, 4)
}
\seealso{
Other Functionals: \code{\link{And}}, \code{\link{Or}};
  \code{\link{Compose}}, \code{\link{Sequence}},
  \code{\link{\%.\%}}; \code{\link{Delay_by}};
  \code{\link{Dot_every}}; \code{\link{FMap}};
  \code{\link{Fail_with}}; \code{\link{Log_to}};
  \code{\link{Maybe}}; \code{\link{Partial}}
}

