% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact.R
\name{compact}
\alias{compact}
\alias{compactChar}
\alias{compactNA}
\title{Filter "empty" elements from a list.}
\usage{
compact(x)

compactChar(x)

compactNA(x)
}
\arguments{
\item{x}{A vector or list.}
}
\value{
A vector or list with empty elements filtered out.
}
\description{
\code{compact} filters \code{NULL} elments from a list; \code{compactChar} filters
empty strings (\code{""}) from a list (or vector); \code{compactNA} filters
\code{NA}s.
}
\examples{
l <- list(a=1, b=NULL, c=NA)
compact(l)
## $a
## [1] 1
## 
## $c
## [1] NA
compactNA(l)
## $a
## [1] 1
## 
## $b
## NULL
}
