% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regexp-utils.R
\name{\%n~\%}
\alias{\%n~\%}
\alias{\%n~f\%}
\title{A pattern counting operator.}
\usage{
x \%n~\% re

x \%n~f\% re
}
\arguments{
\item{x}{A character vector to match \code{re} against.}

\item{re}{A regular expression.}
}
\value{
The count of the elements in
}
\description{
Count how often a pattern occurs in a character vector.
}
\examples{
c("foo", "bar", "baz") \%n~\% "^b"
}
\seealso{
Other Operators: \code{\link{\%<~\%}},
  \code{\link{\%<~f\%}}; \code{\link{\%~\%}},
  \code{\link{\%~f\%}}

Other Operators: \code{\link{\%<~\%}},
  \code{\link{\%<~f\%}}; \code{\link{\%~\%}},
  \code{\link{\%~f\%}}
}

