% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{createPackage}
\alias{createPackage}
\title{Create a new package}
\usage{
createPackage(name, path = getOption("gsmisc.devel"), use_test = TRUE,
  use_doc = TRUE, use_rcpp = FALSE, use_travis = FALSE,
  use_vignette = FALSE)
}
\arguments{
\item{name}{Name of the package.}

\item{path}{Location of the package.}

\item{use_test}{Add testing infrastructure to the package.}

\item{use_doc}{Add a roxygen template for package documentation.}

\item{use_rcpp}{Add infrastructure for using \code{Rcpp}.}

\item{use_travis}{Add a basic travis template.}

\item{use_vignette}{Add a vignette template to the project.}
}
\value{
Called for its side effects. Opens a new instance of Rstudio in
  the newly created package directory.
}
\description{
\code{createPackage} is a simple wrapper around \code{\link[devtools]{create}}.
}
\examples{
\dontrun{
   createPackage("utils", use_rcpp = TRUE)  
}

}

