% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{Delay_by}
\alias{Delay_by}
\title{Delay function call}
\usage{
Delay_by(delay, fn)
}
\arguments{
\item{delay}{delay in seconds.}

\item{fn}{function to call}
}
\description{
Delay function call
}
\seealso{
Other Functionals: \code{\link{And}}, \code{\link{Call}},
  \code{\link{Compose}}, \code{\link{Dot_every}},
  \code{\link{FMap}}, \code{\link{Fail_with}},
  \code{\link{Log_to}}, \code{\link{Maybe}},
  \code{\link{Partial}}
}

