% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexp-utils.R
\name{\%<~\%}
\alias{\%<~\%}
\alias{\%<~f\%}
\title{A pattern extraction operator.}
\usage{
x \%<~\% re

x \%<~f\% re
}
\arguments{
\item{x}{A character vector to match \code{re} against.}

\item{re}{A regular expression.}
}
\value{
The matching elements in \code{x}.
}
\description{
Extract matching elements from a character vector.
}
\seealso{
Other Operators: \code{\link{\%n~\%}}, \code{\link{\%~\%}}

Other Operators: \code{\link{\%n~\%}}, \code{\link{\%~\%}}
}
