% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{\%~\%}
\alias{\%~\%}
\title{A simple pattern matching operator.}
\usage{
x \%~\% re
}
\arguments{
\item{x}{A character vector.}

\item{pattern}{A regular expression.}
}
\value{
Matching elements in \code{target}
}
\description{
A simple pattern matching operator.
}

