% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ht}
\alias{ht}
\title{Show head and tail of a data structure.}
\usage{
ht(x, n = 6)
}
\arguments{
\item{x}{An object.}

\item{n}{How many elements to print.}
}
\value{
Only called for its side effect.
}
\description{
Prints head and tail of avector, list, matrix, or data.frame.
}
\seealso{
\code{\link{head}}, \code{link{tail}}.
}

