% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{is.empty}
\alias{all_empty}
\alias{are_empty}
\alias{is.empty}
\title{Is an object empty?}
\usage{
is.empty(x)

are_empty(x)

all_empty(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
A logical.
}
\description{
Which elements in a list are empty?
Are all elements in a list empty?
}
\examples{
is.empty(NULL)
is.empty(numeric())
is.empty(list())
is.empty("")
are_empty(list(1, NULL, 3, NA))
all_empty(NULL)
all_empty(list(NULL, NULL, character()))
}
\seealso{
Other Tests: \code{\link{all_scalar}},
  \code{\link{are_scalar}}, \code{\link{is.scalar}},
  \code{\link{is.string}}; \code{\link{are_false}};
  \code{\link{are_null}}; \code{\link{are_true}};
  \code{\link{has_command}}; \code{\link{is.installed}}
}

