% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{Log_to}
\alias{Log_to}
\title{Log a time stamp and a message to file everytime a function is run}
\usage{
Log_to(path, message = "", fn)
}
\arguments{
\item{path}{path to log file}

\item{message}{logging message}

\item{fn}{function}
}
\description{
Log a time stamp and a message to file everytime a function is run
}
\seealso{
Other Functionals: \code{\link{And}}, \code{\link{Call}},
  \code{\link{Compose}}, \code{\link{Delay_by}},
  \code{\link{Dot_every}}, \code{\link{FMap}},
  \code{\link{Fail_with}}, \code{\link{Maybe}},
  \code{\link{Partial}}
}

