% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{Partial}
\alias{Partial}
\title{Partial function application.}
\usage{
Partial(fn, ..., .env = parent.frame())
}
\arguments{
\item{fn}{Function to apply partially.}

\item{.env}{the environment of the created function. Defaults to
\code{\link{parent.frame}}.}

\item{\dots}{Arguments that should be applied to \code{fn}}
}
\description{
Partial function application.
}

