% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{Partial}
\alias{Partial}
\title{Partial function application.}
\usage{
Partial(fn, ..., .env = parent.frame())
}
\arguments{
\item{fn}{Function to apply partially.}

\item{.env}{the environment of the created function. Defaults to
\code{\link{parent.frame}}.}

\item{\dots}{Named arguments that should be applied to \code{fn}}
}
\description{
Modify a function by pre-applying some of the arguments.
}
\examples{
mean1 <- Partial(mean, na.rm = TRUE)
mean1(c(1,2,3,NA,4))
}

