% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-utils.R
\name{trim}
\alias{trim}
\title{Trim elements}
\usage{
trim(x, which = c("both", "left", "right"))
}
\arguments{
\item{x}{a character vector}

\item{which}{a character string specifying whether to remove both
    leading and trailing whitespace (default), or only leading
    (\code{"left"}) or trailing (\code{"right"}).  Can be abbreviated.}
}
\value{
A vector with trimmed strings
}
\description{
Trim elements
}
\examples{
x <- " abc\\n\\t" 
trim(x)
}
\seealso{
Other string-utils: \code{\link{blanks}},
  \code{\link{dup}}, \code{\link{pad}},
  \code{\link{replace_ext}}, \code{\link{split_path}},
  \code{\link{strip_ext}}, \code{\link{strsplitN}},
  \code{\link{wrap}}
}

