% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/install.R
\name{update_github}
\alias{update_github}
\title{A quick installer for my R packages on GitHub}
\usage{
update_github()
}
\description{
Prompts for which of the following packages you want to install:
\code{rmisc}, \code{Rentrez}, \code{biofiles}, \code{blastr},
\code{ncbi}.
}
\seealso{
Other Installers: \code{\link{install_packages}},
  \code{\link{update_packages}}
}

