% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/string-utils.R
\name{wrap}
\alias{wrap}
\title{Wrap elements.}
\usage{
wrap(x, wrap = "\\"")
}
\arguments{
\item{x}{A vector.}

\item{wrap}{A string to wrap around vector elements.}
}
\value{
A vector with wrapped strings.
}
\description{
Wrap elements.
}
\examples{
wrap(c("abc", "xyz"), "---")
}
\seealso{
Other string-utils: \code{\link{blanks}};
  \code{\link{dup}}; \code{\link{pad}};
  \code{\link{replace_ext}}; \code{\link{split_path}};
  \code{\link{strip_ext}}; \code{\link{strsplitN}};
  \code{\link{trim}}
}

